[Frontend | Refactor]: Refactored pages views folder and files structure
