+++ frontend/src/views/Editor/Editor.module.scss
